refactor: extract RequestListHeader
